+++ ui/widgets/__init__.py
@@ -1,0 +1,8 @@
+# ui/widgets/__init__.py
+
+"""
+Application widget helper package.
+
+Contains setup and behaviour of important widgets in the application
+such as the log and progressbar.
+"""